requirements.txt
fastapi~=0.68.2
pytest~=6.2.5
python-dotenv~=0.19.1
uvicorn~=0.15.0
